build: enable /W4 on MSVC; remove dead accessors and fix the defects it surfaced
MSVC was building at its DEFAULT warning level -- only /utf-8 was set, while
Clang/GCC got -Wall -Wextra. The CMakeLists comment even claimed the flags
"mirror the MSVC intent", which was not true. Unused locals, shadowed
declarations and narrowing conversions were therefore invisible on Windows
while the macOS/Raspberry Pi builds saw them: a defect found on one host and
not the other is a portability trap.

/W4 is now set alongside /utf-8, with /wd4996 to silence MSVC's nagging about
getenv/fopen/sscanf -- those are portable Standard C and the suggested
replacements (_dupenv_s, fopen_s, ...) are MSVC-only, so it is advice this
project cannot take. -Werror//WX stay off, per existing policy.

Defects it surfaced in src/, now fixed:
  - hbf1xv_machine.cpp: an `else if` re-declared `cart`, shadowing the `if`
    branch's local. Renamed to cart_slot1/cart_slot2.
  - sdl3_main.cpp: `exe_dir` was re-declared in the settings-persistence
    block, shadowing the one resolved for project-root discovery and calling
    exe_directory() a second time. Now reuses the outer value.
  - cpm_bdos_harness.cpp: `kLoadBase + image.size()` promoted to size_t and
    then narrowed implicitly; both operands are now cast explicitly.

Dead code removed -- ten inline accessors/mutators with ZERO callers anywhere
in src/ or tests/ (verified by whole-codebase identifier frequency):
  psg_ym2149: audio_sync_enabled, last_sync_cycle
  click_dac:  current_level, pending_edges
  rom_asset_loader: clear_diagnostics, set_root
  audio_pacer: silence_samples_pushed
  sdl3_app:   input_mapper, quit_requested, request_stream_toggle
(request_stream_toggle was a menu seam that was never wired -- the F10 hotkey
calls on_stream_toggle_hotkey() directly.)

src/ now compiles CLEAN at /W4. The 83 remaining warnings are all in tests/
(narrowing in fixtures, discarded [[nodiscard]] returns, an unused parameter,
one shadowed local) and are catalogued for a follow-up pass.

Also audited: no orphaned headers, no sources missing from the build, no
duplicate definitions, no stray TODO/FIXME debt.

Gates: clean bootstrap build; ctest -LE slow_cpu_sweep 273/273, with both
game-asset regression oracles running for real.

+++ src/machine/cpm_bdos_harness.cpp
@@ -24,7 +24,11 @@ CpmBdosHarness::LoadResult CpmBdosHarness::load_com(Hbf1xvMachine& machine, std:
     // when the load is about to be refused.
     machine.map_flat_ram();
 
-    const std::uint32_t end_address = static_cast<std::uint32_t>(kLoadBase) + image.size();
+    // Both operands are narrowed EXPLICITLY: `kLoadBase + image.size()` would
+    // promote to std::size_t (64-bit) and then narrow implicitly on assignment.
+    // image.size() is bounded by the 64 KB guard below, so the cast is lossless.
+    const std::uint32_t end_address =
+        static_cast<std::uint32_t>(kLoadBase) + static_cast<std::uint32_t>(image.size());
     if (end_address >= top_of_memory) {
         // Defensive, generic safety guard (not a magic number specific to
         // any one CP/M program): refuse to load rather than risk silently
